Changes requested for PR.

+++ src/main/java/net/dries007/tfc/objects/blocks/agriculture/BlockFruitTreeTrunk.java
@@ -290,7 +290,7 @@ public class BlockFruitTreeTrunk extends Block
     public void neighborChanged(IBlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos)
     {
         super.neighborChanged(state, worldIn, pos, blockIn, fromPos);
-        if (!BlocksTFC.isGrowableSoil(worldIn.getBlockState(pos.down())))
+        if (!(worldIn.getBlockState(pos.down()).getBlock() instanceof BlockFruitTreeTrunk) && !BlocksTFC.isGrowableSoil(worldIn.getBlockState(pos.down())))
         {
             worldIn.destroyBlock(pos, false);
         }
